requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
ctranslate2==4.6.0
diskcache==5.6.3
edge-tts==7.0.2
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
keyboard==0.13.5
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
numpy==2.3.1
onnxruntime==1.22.1
packaging==25.0
propcache==0.3.2
protobuf==6.31.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pypiwin32==223
pyreadline3==3.5.4
pywin32==311
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
simpleaudio==1.0.4
sniffio==1.3.1
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
